% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_assign.r
\name{assignPars}
\alias{assignPars}
\title{Assign Values to Parameters}
\value{
\code{NULL} (invisible). The assigned numeric data are stored in the
  object and can be accessed by the \code{\link{queryPars}} method.
}
\description{
Assign values to parameters of a \code{\link{rodeo}}-based model.
}
\examples{
# see the example for method 'assignVars' which behaves in the same way
}
\author{
\email{david.kneis@tu-dresden.de}
}
\seealso{
Other \code{\link{rodeo}} class methods
}

