% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_assign.r
\name{assignVars}
\alias{assignVars}
\title{Assign Values to State Variables}
\arguments{
\item{x}{A matrix, vector, or data frame holding the data to be assigned.
The appropriate type of input depends on the value of \code{tabular}. See
details below.}

\item{tabular}{If set to \code{FALSE}, then \code{x} must be a numeric matrix
(or vector) holding data for \emph{all} items and sections. If \code{TRUE},
then \code{x} must be a data frame with 3 columns 'name', 'section', and
'value'. Use this to assign data to selected items and/or sections only.}

\item{check}{Logical. By default, several checks are carried out on the
passed \code{x}. This can be turned off by setting \code{check} to
\code{FALSE}. May be used to avoid unnecessary checks in repeated calls.}
}
\value{
\code{NULL} (invisible). The assigned numeric data are stored in the
  object and can be accessed by the \code{\link{queryVars}} method.
}
\description{
Assign values to state variables of a \code{\link{rodeo}}-based model.
}
\note{
If \code{tabular} is \code{FALSE}, the matrix passed as \code{x} must
  have column names correspond to the names of state variables. The number of
  rows is expect to match the object's number of spatial sections. In the
  common case of a single spatial section, \code{x} can also be a named
  vector instead of a single-row matrix.
}
\examples{
data(exampleIdentifiers, exampleProcesses, exampleStoichiometry)
model <- rodeo$new(
  vars=subset(exampleIdentifiers, type=="v"),
  pars=subset(exampleIdentifiers, type=="p"),
  funs=subset(exampleIdentifiers, type=="f"),
  pros=exampleProcesses, stoi=exampleStoichiometry,
  size=2
)
# Set all values at a time by passing a matrix
x <- cbind(c_z=c(1,1), c_do=c(9,9), v=c(1e6,1e6))
model$assignVars(x)
print(model$queryVars(asMatrix=TRUE))
# Set selected values by passing a data frame
x <- data.frame(name=c("c_z","c_do"), section=c(1,2), value=c(2,11),
  stringsAsFactors=FALSE)
model$assignVars(x, tabular=TRUE)
print(model$queryVars(asMatrix=TRUE))
}
\author{
\email{david.kneis@tu-dresden.de}
}
\seealso{
Other \code{\link{rodeo}} class methods
}

