% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_ModelBased.R
\name{AutoEncoder_H2O}
\alias{AutoEncoder_H2O}
\title{AutoEncoder_H2O}
\usage{
AutoEncoder_H2O(
  RunMode = "train",
  ArgsList = NULL,
  TrainData. = NULL,
  ValidationData. = NULL,
  TestData. = NULL,
  ScoringData. = NULL,
  Pause = 0L
)
}
\arguments{
\item{RunMode}{'train' will run in train mode. Supply any other character to run scoring mode. Must supply a character}

\item{ArgsList}{ArgsList_FEE}

\item{TrainData.}{data}

\item{ValidationData.}{data}

\item{TestData.}{data}

\item{ScoringData.}{data}

\item{Pause}{0L Sys.sleep(Pause)}
}
\description{
Utilize an H2O autoencoder to provide dimensionality reduction and anomaly detection
}
\seealso{
Other Feature Engineering - Model Based: 
\code{\link{Clustering_H2O}()},
\code{\link{IsolationForest_H2O}()},
\code{\link{Word2Vec_H2O}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering - Model Based}
