% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_CrossRowOperations.R
\name{AutoLagRollMode}
\alias{AutoLagRollMode}
\title{AutoLagRollMode}
\usage{
AutoLagRollMode(
  data,
  Lags = 1,
  ModePeriods = 0,
  Targets = NULL,
  GroupingVars = NULL,
  SortDateName = NULL,
  WindowingLag = 0,
  Type = c("Lag"),
  SimpleImpute = TRUE,
  Debug = FALSE
)
}
\arguments{
\item{data}{A data.table you want to run the function on}

\item{Lags}{A numeric vector of the specific lags you want to have generated. You must include 1 if WindowingLag = 1.}

\item{ModePeriods}{A numberic vector of window sizes}

\item{Targets}{A character vector of the column names for the reference column in which you will build your lags and rolling stats}

\item{GroupingVars}{A character vector of categorical variable names you will build your lags and rolling stats by}

\item{SortDateName}{The column name of your date column used to sort events over time}

\item{WindowingLag}{Set to 0 to build rolling stats off of target columns directly or set to 1 to build the rolling stats off of the lag-1 target}

\item{Type}{List either "Lag" if you want features built on historical values or "Lead" if you want features built on future values}

\item{SimpleImpute}{Set to TRUE for factor level imputation of "0" and numeric imputation of -1}

\item{Debug}{= FALSE}
}
\value{
data.table of original data plus created lags, rolling stats, and time between event lags and rolling stats
}
\description{
Create lags and rolling modes for categorical variables.
}
\examples{
\dontrun{
# NO GROUPING CASE: Create fake Panel Data----
Count <- 1L
for(Level in LETTERS) {
  datatemp <- AutoQuant::FakeDataGenerator(
    Correlation = 0.75,
    N = 25000L,
    ID = 0L,
    ZIP = 0L,
    FactorCount = 2L,
    AddDate = TRUE,
    Classification = FALSE,
    MultiClass = FALSE)
  datatemp[, Factor1 := eval(Level)]
  if(Count == 1L) {
    data <- data.table::copy(datatemp)
  } else {
    data <- data.table::rbindlist(
      list(data, data.table::copy(datatemp)))
  }
  Count <- Count + 1L
}

# NO GROUPING CASE: Create rolling modes for categorical features
data <- AutoQuant::AutoLagRollMode(
  data,
  Lags           = seq(1,5,1),
  ModePeriods    = seq(2,5,1),
  Targets        = c("Factor_1"),
  GroupingVars   = NULL,
  SortDateName   = "DateTime",
  WindowingLag   = 1,
  Type           = "Lag",
  SimpleImpute   = TRUE)

# GROUPING CASE: Create fake Panel Data----
Count <- 1L
for(Level in LETTERS) {
  datatemp <- AutoQuant::FakeDataGenerator(
    Correlation = 0.75,
    N = 25000L,
    ID = 0L,
    ZIP = 0L,
    FactorCount = 2L,
    AddDate = TRUE,
    Classification = FALSE,
    MultiClass = FALSE)
  datatemp[, Factor1 := eval(Level)]
  if(Count == 1L) {
    data <- data.table::copy(datatemp)
  } else {
    data <- data.table::rbindlist(
      list(data, data.table::copy(datatemp)))
  }
  Count <- Count + 1L
}

# GROUPING CASE: Create rolling modes for categorical features
data <- AutoQuant::AutoLagRollMode(
  data,
  Lags           = seq(1,5,1),
  ModePeriods    = seq(2,5,1),
  Targets        = c("Factor_1"),
  GroupingVars   = "Factor_2",
  SortDateName   = "DateTime",
  WindowingLag   = 1,
  Type           = "Lag",
  SimpleImpute   = TRUE)
}
}
\seealso{
Other Feature Engineering: 
\code{\link{Apply_Asinh}()},
\code{\link{Apply_Asin}()},
\code{\link{Apply_BoxCox}()},
\code{\link{Apply_LogPlus1}()},
\code{\link{Apply_Logit}()},
\code{\link{Apply_Log}()},
\code{\link{Apply_Sqrt}()},
\code{\link{Apply_YeoJohnson}()},
\code{\link{AutoDataPartition}()},
\code{\link{AutoDiffLagN}()},
\code{\link{AutoInteraction}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{AutoWord2VecScoring}()},
\code{\link{CategoricalEncoding}()},
\code{\link{CreateCalendarVariables}()},
\code{\link{CreateHolidayVariables}()},
\code{\link{DT_GDL_Feature_Engineering}()},
\code{\link{DummifyDT}()},
\code{\link{Estimate_BoxCox_Lambda}()},
\code{\link{Estimate_YeoJohnson_Lambda}()},
\code{\link{H2OAutoencoderScoring}()},
\code{\link{H2OAutoencoder}()},
\code{\link{Interact}()},
\code{\link{InvApply_Asinh}()},
\code{\link{InvApply_Asin}()},
\code{\link{InvApply_BoxCox}()},
\code{\link{InvApply_LogPlus1}()},
\code{\link{InvApply_Logit}()},
\code{\link{InvApply_Log}()},
\code{\link{InvApply_Sqrt}()},
\code{\link{InvApply_YeoJohnson}()},
\code{\link{ModelDataPrep}()},
\code{\link{Partial_DT_GDL_Feature_Engineering2}()},
\code{\link{Partial_DT_GDL_Feature_Engineering}()},
\code{\link{PercRankScoring}()},
\code{\link{PercRank}()},
\code{\link{StandardizeScoring}()},
\code{\link{Standardize}()},
\code{\link{Test_Asinh}()},
\code{\link{Test_Asin}()},
\code{\link{Test_BoxCox}()},
\code{\link{Test_Identity}()},
\code{\link{Test_LogPlus1}()},
\code{\link{Test_Logit}()},
\code{\link{Test_Log}()},
\code{\link{Test_Sqrt}()},
\code{\link{Test_YeoJohnson}()},
\code{\link{TimeSeriesFillRoll}()},
\code{\link{TimeSeriesFill}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
