% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_ModelBased.R
\name{Clustering_H2O}
\alias{Clustering_H2O}
\title{Clustering_H2O}
\usage{
Clustering_H2O(
  ArgsList = ArgsList_FEE,
  TrainData. = NULL,
  ValidationData. = NULL,
  TestData. = NULL,
  ScoringData. = NULL
)
}
\arguments{
\item{ArgsList}{ArgsList_FEE}

\item{TrainData.}{data}

\item{ValidationData.}{data}

\item{TestData.}{data}

\item{ScoringData.}{data}
}
\description{
Utilize an H2O autoencoder and kmeans to create a cluster id column
}
\seealso{
Other Feature Engineering - Model Based: 
\code{\link{AutoEncoder_H2O}()},
\code{\link{IsolationForest_H2O}()},
\code{\link{Word2Vec_H2O}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering - Model Based}
