% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_NumericTypes.R
\name{CreateInteractions}
\alias{CreateInteractions}
\title{CreateInteractions}
\usage{
CreateInteractions(
  data = NULL,
  RunMode = "train",
  ArgsList = ArgsList,
  SkipCols = NULL
)
}
\arguments{
\item{data}{Source data}

\item{RunMode}{'train' or 'score'}

\item{ArgsList}{ArgsList_FFE}

\item{SkipCols}{Vector of column names to remove from data}
}
\value{
A list containing the data and the ArgsList
}
\description{
Create interaction variables
}
\examples{
\dontrun{
Output <- AutoQuant:::CreateInteractions(
  data = data,
  RunMode = "train",
  ArgsList = ArgsList_FE,
  SkipCols = NULL)
data <- Output$data
ArgsList <- Output$ArgsList
}

}
\seealso{
Other Feature Engineering - Numeric Types: 
\code{\link{DiffLagN}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering - Numeric Types}
