% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_CrossRowOperations.R
\name{DiffDT}
\alias{DiffDT}
\title{DiffDT}
\usage{
DiffDT(data, x, NLag1, NLag2, Type = "numeric")
}
\arguments{
\item{data}{Source data}

\item{x}{Column name}

\item{NLag1}{Numeric}

\item{NLag2}{Numeric}

\item{Type}{Choose from 'numeric' or 'date'}
}
\description{
Difference a column in a data.table
}
\seealso{
Other Misc: 
\code{\link{LB}()}
}
\author{
Adrian Antico
}
\concept{Misc}
