% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_CrossRowOperations.R
\name{DiffLagN}
\alias{DiffLagN}
\title{DiffLagN}
\usage{
DiffLagN(
  data = NULL,
  RunMode = "train",
  ArgsList = ArgsList_FFE,
  SkipCols = NULL,
  N1 = 0,
  N2 = 1,
  RunNumber = 1,
  RemoveNAs = FALSE
)
}
\arguments{
\item{data}{Source data}

\item{RunMode}{'train' or 'score'}

\item{ArgsList}{ArgsList_FFE}

\item{SkipCols}{Vector of column names to remove from data}

\item{N1}{Lookback for time period 1}

\item{N2}{Lookback for time period 2}

\item{RunNumber}{Iteration number when running multiple times}

\item{RemoveNAs}{Remove NULL values created by lags}
}
\value{
A list containing the data and the ArgsList
}
\description{
Create differences for numeric and date variables
}
\examples{
\dontrun{
Output <- AutoQuant:::DiffLagN(
  data = data,
  RunMode = "train",
  ArgsList = ArgsList_FE,
  SkipCols = NULL,
  N1 = 0,
  N2 = 1,
  RunNumber = 1,
  RemoveNAs = FALSE)
data <- Output$data
ArgsList_FE <- Output$ArgsList
}

}
\seealso{
Other Feature Engineering - Numeric Types: 
\code{\link{CreateInteractions}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering - Numeric Types}
