% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_CharacterTypes.R
\name{DummyVariables}
\alias{DummyVariables}
\title{DummyVariables}
\usage{
DummyVariables(
  data,
  RunMode = "train",
  ArgsList = NULL,
  SkipCols = NULL,
  KeepCharCols = TRUE
)
}
\arguments{
\item{data}{Source data}

\item{RunMode}{'train' or 'score'}

\item{ArgsList}{ArgsList_FFE}

\item{SkipCols}{Vector of column names to remove from data}
}
\value{
A list containing the data and the ArgsList
}
\description{
Create dummy variables for categorical variables. You can select the max amount of levels to return per feature.
}
\examples{
\dontrun{
Output <- AutoQuant::DummyVariables(
  data = data,
  RunMode = "train",
  ArgsList = ArgsList_FE,
  SkipCols = NULL)
data <- Output$data
ArgsList_FE <- Output$ArgsList
}

}
\author{
Adrian Antico
}
\concept{Feature Engineering - Character Types}
