% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_CharacterTypes.R
\name{EncodeCharacterVariables}
\alias{EncodeCharacterVariables}
\title{EncodeCharacterVariables}
\usage{
EncodeCharacterVariables(
  RunMode = "train",
  ModelType = "classification",
  TrainData = NULL,
  ValidationData = NULL,
  TestData = NULL,
  TargetVariableName = NULL,
  CategoricalVariableNames = NULL,
  EncodeMethod = NULL,
  KeepCategoricalVariables = FALSE,
  ReturnMetaData = FALSE,
  MetaDataPath = NULL,
  MetaDataList = NULL,
  ImputeMissingValue = 0,
  Debug = FALSE
)
}
\arguments{
\item{RunMode}{'train' or 'score'}

\item{ModelType}{'classification', 'regression', 'multiclass'}

\item{TrainData}{Must supply data.table}

\item{ValidationData}{Optional}

\item{TestData}{Optional}

\item{TargetVariableName}{Column name}

\item{CategoricalVariableNames}{Column names}

\item{EncodeMethod}{Choose from 'binary', 'm_estimator', 'credibility', 'woe', 'target_encoding', 'poly_encode', 'backward_difference', 'helmert'}

\item{KeepCategoricalVariables}{Logical}

\item{ReturnMetaData}{Logical}

\item{MetaDataPath}{Supply a directory path or NULL}

\item{MetaDataList}{Supply a metadata list or NULL}

\item{ImputeMissingValue}{Supply a value or leave NULL to handle elsewhere}

\item{Debug}{= FALSE}
}
\description{
EncodeCharacterVariables
}
