% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_CharacterTypes.R
\name{Encoding}
\alias{Encoding}
\title{Encoding}
\usage{
Encoding(
  RunMode = "train",
  ArgsList = NULL,
  TrainData = NULL,
  ValidationData = NULL,
  TestData = NULL,
  ScoringData = NULL
)
}
\arguments{
\item{RunMode}{Passthrough}

\item{ArgsList}{Passthrough}

\item{TrainData}{Passthrough}

\item{ValidationData}{Passthrough}

\item{TestData}{Passthrough}

\item{ScoringData}{Passthrough}
}
\description{
Encoding
}
