% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{exampleIdentifiers}
\alias{exampleIdentifiers}
\title{Declaration of identifiers}
\format{A data frame with the following fields:
  \itemize{
    \item{name : }{Name of the item}
    \item{type : }{'v' for variable, 'p' for paraneter, 'f' for function}
    \item{units : }{Units of the item}
    \item{description : }{Short description (text)}
  }}
\description{
Declaration of items (i.e. identifiers) appearing in the model's
mathematical expressions.
}

