% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{exampleStoichiometry}
\alias{exampleStoichiometry}
\title{Specification of Stoichiometry}
\format{A data frame with the following fields:
  \itemize{
    \item{variable : }{Name of the state variable}
    \item{process : }{Name of the process}
    \item{expression : }{Mathematical expression (as a string)}
  }}
\description{
Definition of the links between simulated processes and state variables.
}

