% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethod_finalize.r
\name{finalize}
\alias{finalize}
\title{Clean-up a \code{rodeo} Object}
\value{
The method is called implicitly for its side effects when a
  \code{\link{rodeo}} object is destroyed.
}
\description{
Clean-up method for objects of the \code{\link{rodeo-class}}.
}
\note{
At present, the method just unloads the object-specific
  shared libraries created with the \code{\link{compile}} or
  \code{\link{initStepper}} methods.
}
\author{
\email{david.kneis@tu-dresden.de}
}

