% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_forcings.r
\name{forcings_clear}
\alias{forcings_clear}
\title{Clear forcing data of a Fortran-based model}
\value{
Returns \code{NULL} invisibly.
}
\description{
Clears the forcings data of a model written in Fortran. This may be
useful when data should be updated through another call to to
\code{\link{forcings_init}} or for the purpose of memory cleanup.
}
\note{
The function takes no arguments and should thus called as
  \code{forcings_clear()}.
}
\seealso{
\code{\link{forcings_init}} to initialize the forcing data
}
\author{
\email{david.kneis@tu-dresden.de}
}
