% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_forcings.r
\name{forcings_init}
\alias{forcings_init}
\title{Initialize forcings of a Fortran-based model}
\value{
Returns \code{NULL} invisibly.
}
\description{
Initializes the forcings of a model written in Fortran based on the
input supplied to \code{forcingFunctions}. If the data cannot be
initialized for any reason, an error is thrown and a hopefully
useful message is provided.
}
\note{
The function takes no arguments and should thus called as
  \code{forcings_init()}.
}
\seealso{
\code{\link{forcings_clear}} to clear the data set by
 \code{forcings_init}
}
\author{
\email{david.kneis@tu-dresden.de}
}
