% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fortranWrapper_deSolve.r
\name{fortranWrapper_deSolve}
\alias{fortranWrapper_deSolve}
\title{Fortran wrapper for deSolve methods}
\usage{
fortranWrapper_deSolve(NLVL = 1)
}
\arguments{
\item{NLVL}{The desired number of spatial levels (boxes). Defaults to 1.}
}
\value{
A character string representing Fortran code. Must be written to
  disk, e.g. using \code{write} prior to compilation.
}
\description{
Creates wrapper code to make generated Fortran code compatible with the
integrators from package \code{deSolve} or methods from \code{rootSolve}.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

