% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_vNext.R
\name{generate_rodeo_feature_engineering_artifacts}
\alias{generate_rodeo_feature_engineering_artifacts}
\title{Generate feature engineering artifacts from a Rodeo vNext run}
\usage{
generate_rodeo_feature_engineering_artifacts(
  data,
  plan = NULL,
  fitted_plan = NULL,
  benchmark_summary = NULL
)
}
\description{
Generate feature engineering artifacts from a Rodeo vNext run
}
