% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrep_vNext.R
\name{generate_rodeo_model_prep_artifacts}
\alias{generate_rodeo_model_prep_artifacts}
\title{Generate Rodeo vNext model-prep artifacts}
\usage{
generate_rodeo_model_prep_artifacts(data, plan = NULL, fitted_plan = NULL)
}
\description{
Generate Rodeo vNext model-prep artifacts
}
