% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_getValues.r
\name{getPars}
\alias{getPars}
\title{Query Values of Parameters}
\arguments{
\item{asMatrix}{Logical. If \code{TRUE}, data are returned as a matrix with
columns named after the parameters and one row per spatial section. If
\code{FALSE}, a vector of the concatenated columns is returned.
In the latter case, element names follow the pattern
'parameter.section', i.e. the section index is appended to the parameter's
name, separated by a period. If there is a single section only (0D models),
the pure parameter names are used, i.e. the extension '.1' is omitted.}
}
\value{
A numeric matrix or vector, depending on the value of \code{asMatrix}.
}
\description{
Query values of parameters of a \code{\link{rodeo}}-based model.
}
\examples{
data(exampleIdentifiers, exampleProcesses, exampleStoichiometry)
model <- rodeo$new(
  vars=subset(exampleIdentifiers, type=="v"),
  pars=subset(exampleIdentifiers, type=="p"),
  funs=subset(exampleIdentifiers, type=="f"),
  pros=exampleProcesses, stoi=exampleStoichiometry,
  size=2
)
x <- c(kd=5.78e-7, h_do=0.5, s_do_z=2.76,
  wind=1, depth=2, temp=20, q_in=1, q_ex=1)
x <- rbind(x, x)
model$setPars(x)
print(model$getPars())
print(model$getPars(asMatrix=TRUE))
}
\author{
\email{david.kneis@tu-dresden.de}
}
\seealso{
The corresponding 'set' method is \code{\link{setPars}}. Use
  \code{\link{getVars}} to query the values of variables rather than parameters.
}

