% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_getValues.r
\name{getVars}
\alias{getVars}
\title{Query Values of State Variables}
\arguments{
\item{asMatrix}{Logical. If \code{TRUE}, data are returned as a matrix with
columns named after the variables and one row per spatial section. If
\code{FALSE}, a vector of the concatenated columns is returned.
In the latter case, element names follow the pattern
'variable.section', i.e. the section index is appended to the variable's
name, separated by a period. If there is a single section only (0D models),
the pure variable names are used, i.e. the extension '.1' is omitted.}
}
\value{
A numeric matrix or vector, depending on the value of \code{asMatrix}.
}
\description{
Query values of state variables of a \code{\link{rodeo}}-based model.
}
\examples{
data(exampleIdentifiers, exampleProcesses, exampleStoichiometry)
model <- rodeo$new(
  vars=subset(exampleIdentifiers, type=="v"),
  pars=subset(exampleIdentifiers, type=="p"),
  funs=subset(exampleIdentifiers, type=="f"),
  pros=exampleProcesses, stoi=exampleStoichiometry,
  size=2
)
model$setVars(cbind(c_z=c(1,1), c_do=c(9,9), v=c(1e6, 1e6)))
print(model$getVars())
print(model$getVars(asMatrix=TRUE))
}
\author{
\email{david.kneis@tu-dresden.de}
}
\seealso{
The corresponding 'set' method is \code{\link{setVars}}. Use
  \code{\link{getPars}} to query the values of parameters rather than variables.
}

