% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_ModelBased.R
\name{IsolationForest_H2O}
\alias{IsolationForest_H2O}
\title{IsolationForest_H2O}
\usage{
IsolationForest_H2O(
  ArgsList = ArgsList,
  TrainData. = NULL,
  ValidationData. = NULL,
  TestData. = NULL,
  ScoringData. = NULL
)
}
\arguments{
\item{ArgsList}{ArgsList}

\item{TrainData.}{data}

\item{ValidationData.}{data}

\item{TestData.}{data}

\item{ScoringData.}{data}
}
\description{
Utilize an H2O isolation forest to provide anomaly detection
}
\seealso{
Other Feature Engineering - Model Based: 
\code{\link{AutoEncoder_H2O}()},
\code{\link{Clustering_H2O}()},
\code{\link{Word2Vec_H2O}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering - Model Based}
