% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_lib.r
\name{libFunc}
\alias{libFunc}
\title{Return name of library function}
\value{
The name of the function to compute derivatives which is contained in
  the library built with \code{\link{compile}}. This name has to be supplied as
  the \code{func} argument of the solver methods
  in \code{\link[deSolve]{deSolve}} or \code{\link[rootSolve]{rootSolve}}.
}
\description{
Return the name of the library function for use with
 \code{\link[deSolve]{deSolve}} or \code{\link[rootSolve]{rootSolve}} methods.
}
\author{
\email{david.kneis@tu-dresden.de}
}

