% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_lib.r
\name{libName}
\alias{libName}
\title{Return library name}
\value{
The base name of the shared library file created with
  \code{\link{compile}} after stripping of the the platform specific
  extension. This name has to be supplied as the \code{dllname} argument of
  the solver methods in \code{\link[deSolve]{deSolve}} or
  \code{\link[rootSolve]{rootSolve}}.
}
\description{
Return the pure name of the shared library for use with
 \code{\link[deSolve]{deSolve}} or \code{\link[rootSolve]{rootSolve}} methods.
}
\author{
\email{david.kneis@tu-dresden.de}
}
