% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_CrossRowOperations.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{vector}
}
\description{
Statistical mode. Only returns the first mode if there are many
}
\author{
Adrian Antico
}
\concept{EDA}
