% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pars}
\alias{pars}
\title{Declaration of Parameters}
\format{
A data frame with the following fields:
  \describe{
    \item{name : }{Name of the parameter}
    \item{units : }{Unit}
    \item{description : }{Short description (text)}
  }
}
\description{
Declaration of parameters of the bacteria growth example model.
}
