% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_DataSets.R
\name{PartitionData}
\alias{PartitionData}
\title{PartitionData}
\usage{
PartitionData(data = NULL, ArgsList = ArgsList)
}
\arguments{
\item{data}{Source data}

\item{ArgsList}{ArgsList}
}
\value{
A list containing the data and the ArgsList
}
\description{
Create data sets for machine learning
}
\examples{
\dontrun{
Output <- AutoQuant:::PartitionData(
  data = data,
  ArgsList = ArgsList)
TrainData <- Output$TrainData
ArgsList <- Output$ArgsList
}

}
\author{
Adrian Antico
}
\concept{Feature Engineering - DataSet}
