% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_NumericTypes.R
\name{PercRank}
\alias{PercRank}
\title{PercRank}
\usage{
PercRank(
  data,
  ColNames,
  GroupVars = NULL,
  Granularity = 0.001,
  ScoreTable = FALSE
)
}
\arguments{
\item{data}{Source data.table}

\item{ColNames}{Character vector of column names}

\item{GroupVars}{Character vector of column names to have percent ranks by the group levels}

\item{Granularity}{Provide a value such that data.table::frank(Variable) * (1 / Granularity) / .N * Granularity. Default is 0.001}

\item{ScoreTable}{= FALSE. Set to TRUE to get the reference values for applying to new data. Pass to scoring version of this function}
}
\description{
Generate percent ranks for multiple variables, by groups if provided, and with a selected granularity
}
\examples{
\dontrun{
data <- data.table::fread(file.choose())
x <- PercRank(data, ColNames = c('Weekly_Sales', 'XREG1'), GroupVars = c('Region','Store','Dept'), Granularity = 0.001, ScoreTable = TRUE)
}

}
\seealso{
Other Feature Engineering: 
\code{\link{AutoDataPartition}()},
\code{\link{AutoDiffLagN}()},
\code{\link{AutoInteraction}()},
\code{\link{AutoLagRollMode}()},
\code{\link{AutoLagRollStatsScoring}()},
\code{\link{AutoLagRollStats}()},
\code{\link{AutoTransformationCreate}()},
\code{\link{AutoTransformationScore}()},
\code{\link{AutoWord2VecModeler}()},
\code{\link{AutoWord2VecScoring}()},
\code{\link{CategoricalEncoding}()},
\code{\link{CreateCalendarVariables}()},
\code{\link{CreateHolidayVariables}()},
\code{\link{DummifyDT}()},
\code{\link{H2OAutoencoderScoring}()},
\code{\link{H2OAutoencoder}()},
\code{\link{ModelDataPrep}()},
\code{\link{PercRankScoring}()},
\code{\link{StandardizeScoring}()},
\code{\link{Standardize}()},
\code{\link{TimeSeriesFillRoll}()},
\code{\link{TimeSeriesFill}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering}
