% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\name{processes}
\alias{processes}
\title{Processes}
\format{A data frame with the following fields:
  \itemize{
    \item{name : }{Name of the process}
    \item{units : }{Units of the rate expression}
    \item{description : }{Short description (text)}
    \item{expression : }{Mathematical expression (as a string)}
  }}
\description{
Definition of simulated processes.
}

