% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationContract_vNext.R
\name{qa_rodeo_transformation_contract}
\alias{qa_rodeo_transformation_contract}
\title{QA for the structured Rodeo transformation contract}
\usage{
qa_rodeo_transformation_contract()
}
\description{
QA for the structured Rodeo transformation contract
}
