% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_query.r
\name{queryPars}
\alias{queryPars}
\title{Query Values of Parameters}
\arguments{
\item{asMatrix}{Logical. If \code{TRUE}, data are returned as a matrix with
columns named after the parameters and one row per spatial section. If
\code{FALSE}, a vector of the concatenated columns is returned. In that
case, element names follow the pattern 'parameter.section', i.e. the
section index is appended to the parameter name, separated by a period."}
}
\value{
A numeric matrix or vector, depending on the value of \code{asMatrix}.
}
\description{
Query values of parameters of a \code{\link{rodeo}}-based model.
}
\examples{
# see example for method 'queryVars'
}
\author{
\email{david.kneis@tu-dresden.de}
}
\seealso{
Other \code{\link{rodeo}} class methods
}

