% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_query.r
\name{queryVars}
\alias{queryVars}
\title{Query Values of State Variables}
\arguments{
\item{asMatrix}{Logical. If \code{TRUE}, data are returned as a matrix with
columns named after the variables and one row per spatial section. If
\code{FALSE}, a vector of the concatenated columns is returned. In that
case, element names follow the pattern 'variable.section', i.e. the
section index is appended to the variable name, separated by a period."}
}
\value{
A numeric matrix or vector, depending on the value of \code{asMatrix}.
}
\description{
Query values of state variables of a \code{\link{rodeo}}-based model.
}
\examples{
data(exampleIdentifiers, exampleProcesses, exampleStoichiometry)
model= new("rodeo",
  vars=subset(exampleIdentifiers, type=="v"),
  pars=subset(exampleIdentifiers, type=="p"),
  funs=subset(exampleIdentifiers, type=="f"),
  pros=exampleProcesses, stoi=exampleStoichiometry,
  sections=2
)
model$assignVars(cbind(c_z=c(1,1), c_do=c(9,9), v=c(1e6, 1e6)))
print(model$queryVars())
print(model$queryVars(asMatrix=TRUE))
}
\author{
\email{david.kneis@tu-dresden.de}
}
\seealso{
Other \code{\link{rodeo}} class methods
}

