% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrep_vNext.R
\name{rodeo_apply_partition_plan}
\alias{rodeo_apply_partition_plan}
\title{Apply a fitted Rodeo vNext partition plan}
\usage{
rodeo_apply_partition_plan(data, fitted_plan, copy_data = TRUE)
}
\description{
Apply a fitted Rodeo vNext partition plan
}
