% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalTransformation_vNext.R
\name{rodeo_apply_temporal_transformation}
\alias{rodeo_apply_temporal_transformation}
\title{Apply a fitted Rodeo temporal transformation}
\usage{
rodeo_apply_temporal_transformation(data, fitted_spec, copy_data = TRUE)
}
\description{
Apply a fitted Rodeo temporal transformation
}
