% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationContract_vNext.R
\name{rodeo_apply_transformation}
\alias{rodeo_apply_transformation}
\title{Apply a fitted Rodeo transformation specification}
\usage{
rodeo_apply_transformation(data, fitted_spec, copy_data = TRUE)
}
\description{
Apply a fitted Rodeo transformation specification
}
