% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrep_vNext.R
\name{rodeo_create_folds}
\alias{rodeo_create_folds}
\title{Create reproducible fold assignments}
\usage{
rodeo_create_folds(
  data,
  k = 5L,
  target_col = NULL,
  group_col = NULL,
  seed = 123L,
  row_id_col = ".row_id",
  fold_col = ".fold_id"
)
}
\description{
Create reproducible fold assignments
}
