% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationContract_vNext.R
\name{rodeo_fit_apply_transformation}
\alias{rodeo_fit_apply_transformation}
\title{Fit and apply a structured Rodeo transformation}
\usage{
rodeo_fit_apply_transformation(data, spec, copy_data = TRUE)
}
\description{
Fit and apply a structured Rodeo transformation
}
