% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_vNext.R
\name{rodeo_fit_feature_plan}
\alias{rodeo_fit_feature_plan}
\title{Fit a Rodeo vNext feature engineering plan}
\usage{
rodeo_fit_feature_plan(data, plan = rodeo_feature_plan())
}
\description{
Fit a Rodeo vNext feature engineering plan
}
