% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPrep_vNext.R
\name{rodeo_fit_partition_plan}
\alias{rodeo_fit_partition_plan}
\title{Fit a Rodeo vNext partition plan}
\usage{
rodeo_fit_partition_plan(data, plan = rodeo_partition_plan())
}
\description{
Fit a Rodeo vNext partition plan
}
