% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalTransformation_vNext.R
\name{rodeo_fit_temporal_transformation}
\alias{rodeo_fit_temporal_transformation}
\title{Fit a Rodeo temporal transformation specification}
\usage{
rodeo_fit_temporal_transformation(data, spec, forecast_origin = NULL)
}
\description{
Fit a Rodeo temporal transformation specification
}
