% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_vNext.R
\name{rodeo_fit_transform_feature_plan}
\alias{rodeo_fit_transform_feature_plan}
\title{Fit and transform a Rodeo vNext feature plan}
\usage{
rodeo_fit_transform_feature_plan(
  data,
  plan = rodeo_feature_plan(),
  copy_data = TRUE
)
}
\description{
Fit and transform a Rodeo vNext feature plan
}
