% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationContract_vNext.R
\name{rodeo_load_transformation}
\alias{rodeo_load_transformation}
\title{Load a fitted Rodeo transformation specification}
\usage{
rodeo_load_transformation(path)
}
\description{
Load a fitted Rodeo transformation specification
}
