% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationContract_vNext.R
\name{rodeo_save_transformation}
\alias{rodeo_save_transformation}
\title{Save a fitted Rodeo transformation specification}
\usage{
rodeo_save_transformation(fitted_spec, path)
}
\description{
Save a fitted Rodeo transformation specification
}
