% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalTransformation_vNext.R
\name{rodeo_temporal_transformation_metadata}
\alias{rodeo_temporal_transformation_metadata}
\title{Return readable temporal transformation metadata}
\usage{
rodeo_temporal_transformation_metadata(fitted_spec, data = NULL)
}
\description{
Return readable temporal transformation metadata
}
