% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_vNext.R
\name{rodeo_transform_feature_plan}
\alias{rodeo_transform_feature_plan}
\title{Transform data with a fitted Rodeo vNext feature plan}
\usage{
rodeo_transform_feature_plan(data, fitted_plan, copy_data = TRUE)
}
\description{
Transform data with a fitted Rodeo vNext feature plan
}
