% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationContract_vNext.R
\name{rodeo_transformation_metadata}
\alias{rodeo_transformation_metadata}
\title{Read fitted Rodeo transformation metadata}
\usage{
rodeo_transformation_metadata(fitted_spec, data = NULL)
}
\description{
Read fitted Rodeo transformation metadata
}
