% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalTransformation_vNext.R
\name{rodeo_validate_temporal_schema}
\alias{rodeo_validate_temporal_schema}
\title{Validate data against a fitted temporal transformation schema}
\usage{
rodeo_validate_temporal_schema(data, fitted_spec, future_data = NULL)
}
\description{
Validate data against a fitted temporal transformation schema
}
