% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationContract_vNext.R
\name{rodeo_validate_transformation_schema}
\alias{rodeo_validate_transformation_schema}
\title{Validate data against a fitted transformation schema}
\usage{
rodeo_validate_transformation_schema(data, fitted_spec)
}
\description{
Validate data against a fitted transformation schema
}
