% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_CrossRowOperations.R
\name{TimeSeriesFeatures}
\alias{TimeSeriesFeatures}
\title{TimeSeriesFeatures}
\usage{
TimeSeriesFeatures(
  data = NULL,
  RunMode = "train",
  ArgsList = NULL,
  KeepRowsColumnName = NULL,
  KeepRowsGroupID = NULL,
  SkipCols = NULL,
  DebugMode = FALSE
)
}
\arguments{
\item{data}{Source data}

\item{RunMode}{'train' or 'score'}

\item{ArgsList}{ArgsList_FFE}

\item{KeepRowsColumnName}{'scoring' mode. Name of column where subset values reside}

\item{KeepRowsGroupID}{id values to use for subsetting for records to be scored}

\item{SkipCols}{Vector of column names to remove from data}

\item{DebugMode}{Logical}
}
\value{
A list containing the data and the ArgsList
}
\description{
Create lags and rolling stats
}
\examples{
\dontrun{
Output <- Rodeo:::TimeSeriesFeatures(
  data = data,
  RunMode = "train",
  ArgsList = NULL,
  KeepRowsColumnName = NULL,
  KeepRowsGroupID = NULL,
  SkipCols = NULL,
  DebugMode = FALSE)
data <- Output$data
ArgsList <- Output$ArgsList
}

}
\author{
Adrian Antico
}
\concept{Feature Engineering - Cross Row Operations}
