% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\name{toymodel_vars}
\alias{toymodel_vars}
\title{Toy model variables}
\format{A data frame with the following columns:
  \itemize{
    \item{name : }{Name of the variable}
    \item{value : }{Value}
    \item{unit : }{Unit of the value}
    \item{description : }{Short description of the variable's meaning}
  }}
\description{
Sample values for the state variables of \code{toymodel}
}

