% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{vars}
\alias{vars}
\title{Declaration of Variables}
\format{
A data frame with the following fields:
  \describe{
    \item{name : }{Name of the variable}
    \item{units : }{Unit}
    \item{description : }{Short description (text)}
  }
}
\description{
Declaration of variables of the bacteria growth example model.
}
