% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_CalendarTypes.R
\name{weekdays_in_month}
\alias{weekdays_in_month}
\title{Add Number of Weekdays in Month}
\usage{
weekdays_in_month(data, date_col, new_col = "weekdays_in_month")
}
\arguments{
\item{date_col}{The name of the date column (unquoted or quoted).}

\item{new_col}{Name of the new column to create. Default is `"weekdays_in_month"`.}

\item{DT}{A data.table.}
}
\value{
Returns the modified data.table with one additional column containing the
weekday count for each observation's corresponding month.
}
\description{
Given a data.table containing a date column, this function adds a new column
containing the number of weekdays (Monday–Friday) in that date's month.
}
\examples{
library(data.table)
DT <- data.table(date = as.Date(c("2024-01-05", "2024-01-20")))
add_weekdays_in_month(DT, date)

}
