% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEngineering_ModelBased.R
\name{Word2Vec_H2O}
\alias{Word2Vec_H2O}
\title{Word2Vec_H2O}
\usage{
Word2Vec_H2O(
  TrainData. = NULL,
  ValidationData. = NULL,
  TestData. = NULL,
  ScoringData. = NULL,
  ArgsList = ArgsList,
  RunMode = "train",
  SkipCols = NULL
)
}
\arguments{
\item{TrainData.}{Source data}

\item{ValidationData.}{Source data}

\item{TestData.}{Source data}

\item{ScoringData.}{Source data}

\item{ArgsList}{ArgsList_FFE}

\item{RunMode}{'train' or 'score'}

\item{SkipCols}{Colnames to skip over}
}
\description{
Word2Vec modeling and scoring. Features are automatically attached to source data sets
}
\seealso{
Other Feature Engineering - Model Based: 
\code{\link{AutoEncoder_H2O}()},
\code{\link{Clustering_H2O}()},
\code{\link{IsolationForest_H2O}()}
}
\author{
Adrian Antico
}
\concept{Feature Engineering - Model Based}
