% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethod_compile.r
\name{compile}
\alias{compile}
\title{Create Fortran Library}
\arguments{
\item{sources}{Name(s) of fortran source file(s) where a module with the
fixed name 'functions' is implemented. This module must contain all
user-defined functions referenced in process rates or
stoichiometric factors. Can be the name of a single file or, if the
Fortran code is split over several dependent files, a vector of file names.}

\item{target}{Name of a target 'environment' for which the library is
compiled. Currently, 'deSolve' is the only supported value.}
}
\value{
A vector of character strings with named elements as follows:
\itemize{
  \item{\code{libFile}} File path of the created library. Needs to be
    passed to, e.g., \code{\link[base]{dyn.load}}. The library is generally
    created in the folder returned by \code{\link[base]{tempdir}}.
  \item{\code{libName}} The pure library name, which is the base name of
    \code{libFile} with the platform specific extension stripped. This name
    has to be supplied as the \code{dllname} argument of the solver methods
    in \code{\link[deSolve]{deSolve}} or \code{\link[rootSolve]{rootSolve}}.
  \item{\code{libFunc}} Name of the method contained in the built library
    which computes the derivatives. This name has to be supplied as the
    \code{func} argument of the solver methods
    in \code{\link[deSolve]{deSolve}} or \code{\link[rootSolve]{rootSolve}}.
}
}
\description{
Create a Fortran library for use with numerical methods from packages
\code{\link[deSolve]{deSolve}} or \code{\link[rootSolve]{rootSolve}}.
}
\examples{
data(vars, pars, funs, pros, stoi)
model <- rodeo$new(vars, pars, funs, pros, stoi, dim=c(1))
# This would trigger compilation assuming that 'functionsCode.f95' contains
# a Fortran implementation of all functions; see vignette for full example
\dontrun{
lib <- model$compile(sources="functionsCode.f95")
}
}
\author{
\email{david.kneis@tu-dresden.de}
}
\seealso{
This method internally calls \code{\link{generate}}.
}

