% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethod_plotStoichiometry.r
\name{plotStoichiometry}
\alias{plotStoichiometry}
\title{Plot Qualitative Stoichiometry Matrix}
\arguments{
\item{level}{A positive integer representing a spatial section. The values of
state variables and parameters must have been set using the
\code{\link{assignVars}} and \code{\link{assignPars}} methods.}

\item{Time.}{The value is ignored in the case of autonomous models.}

\item{cex}{Character expansion factor.}

\item{colPositive}{Color for positive stoichiometric factors.}

\item{colNegative}{Color for negative stoichiometric factors.}

\item{colGrid}{Color of a grid (can be identical to background color).}
}
\description{
Visualizes the stoichiometry matrix using standard plot methods. The sign
of stoichiometric factors is displayed as upward and downward pointing
triangles, optionally colored.
}
\note{
If the stoichiometric factors are mathematical expressions involving
  function references, these functions must be defined in R (even if the
  numerical computations are based on generated Fortran code).
}
\examples{
data(exampleIdentifiers, exampleProcesses, exampleStoichiometry)
model= new("rodeo",
  vars=subset(exampleIdentifiers, type=="v"),
  pars=subset(exampleIdentifiers, type=="p"),
  funs=subset(exampleIdentifiers, type=="f"),
  pros=exampleProcesses, stoi=exampleStoichiometry
)
c_z_in= function(time) {0.1}
c_do_in= function(time) {8.0}
model$assignVars(c(c_z=1, c_do=9.022, v=1.e6))
model$assignPars(c(kd=5.78e-7, h_do=0.5, s_do_z=2.76, wind=1, depth=2,
  temp=20, q_in=1, q_ex=1))
model$plotStoichiometry(section=1, time=0)
}
\author{
\email{david.kneis@tu-dresden.de}
}
\seealso{
See other methods of the \code{\link{rodeo-class}} or
  \code{\link{stoichiometry}} for computing the stoichiometric factors only.
  Alternative options for displaying stoichiometry information are described
  in the package vignette.
}

