% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pros}
\alias{pros}
\title{Declaration of Processes}
\format{A data frame with the following fields:
  \itemize{
    \item{name : }{Name of the process}
    \item{units : }{Unit of the rate expression}
    \item{description : }{Short description (text)}
    \item{expression : }{Mathematical expression (as a string)}
  }}
\description{
Definition of processes of the bacteria growth example model.
}

