% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_rodeoref.r
\docType{class}
\name{rodeo-class}
\alias{rodeo}
\alias{rodeo-class}
\title{A reference class to represent an ODE-based model.}
\description{
A reference class to represent an ODE-based model.
}
\section{Fields}{

\describe{
\item{\code{pros}}{A named vector of character strings holding process rate expressions.}

\item{\code{stox}}{A data frame holding character strings to represent the
stoichiometry factors (colums = variables, rows = processes).}

\item{\code{auxs}}{A named vector of character strings holding auxiliary expressions being referenced
in, e.g., \code{pros} and/or \code{stox}.}

\item{\code{vars}}{A vector of character strings holding the names of state variables.}

\item{\code{pars}}{A vector of character strings holding the names of parameters.}

\item{\code{funs}}{A vector of character strings holding the names of functions.
  appearing in any of the model's expressions, i.e. \code{pros}, \code{stox},
  and \code{auxs}.}
}}
\section{Methods}{

\describe{
\item{\code{generate(name = "derivs", lang = "r")}}{Generate code to compute the variables' derivatives with respect to time.
\bold{Arguments:} \code{name}: A string giving the name for the generated
function. \code{lang}: The language of generated code (currently 'r' or 'f').
\bold{Returns:} The generated function as a character string.}

\item{\code{getAuxs()}}{Returns a named vector of auxiliary expressions referenced in, e.g., \code{stox} or \code{pros}}

\item{\code{getFuns()}}{Returns a vector of function names appearing in \code{auxs}, \code{stox}, or \code{pros}}

\item{\code{getPars()}}{Returns a vector holding the names of parameters}

\item{\code{getPros()}}{Returns a named vector of process rate expressions}

\item{\code{getStox()}}{Returns the stoichiometry matrix as a data frame (columns = variables, rows = processes)}

\item{\code{getVars()}}{Returns a vector holding the names of state variables}

\item{\code{initialize(file, keyPros = "processes", keyStox = "stoichiometry",
  keyAuxs = "auxiliary")}}{Initializes a \code{rodeo} object with json-formatted data from \code{file}}

\item{\code{plot(pars, vars)}}{Creates a plot of the stoichiometry matrix for the parameter and
variable values in \code{pars} and \code{vars}, respectively}

\item{\code{print()}}{Prints all fields of a \code{rodeo} object}

\item{\code{publish(file, tbl, comment = "\%", colsep = "&", endline = "\\\\\\\\",
  equal = " &= ", startCell = "$ ", endCell = " $")}}{Experimental: Exports model equations to a file for inclusion in e.g. LaTex documents}

\item{\code{sortLikePars(v)}}{Sorts a named vector \code{v} according to the order of parameters}

\item{\code{sortLikeVars(v)}}{Sorts a named vector \code{v} according to the order of state variables}
}}
\examples{
# Create file with model definition in json format
data(toymodel)
tf= tempfile()
write(x=toymodel, file=tf)

# Instantiate object and import model from file
x=new("rodeo", file=tf)

# Display the entire model
x$show()   # built-in method
x$print()

# Define functions referenced in the model
DOSAT= function(t) {14.652 - 0.41022*t + 0.007991*t^2 - 0.000077774*t^3}
k2= function(u, d) {(0.728*sqrt(u) - 0.317*u + 0.0372*u^2) / d / 86400}
# Assign values to state variables and parameters
# then visualize the stoichiometry matrix
data(toymodel_vars)
vars= x$sortLikeVars(setNames(toymodel_vars$value, toymodel_vars$name))
data(toymodel_pars)
pars= x$sortLikePars(setNames(toymodel_pars$value, toymodel_pars$name))

x$plot(vars=vars, pars=pars)
}

