% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods_setValues.r
\name{setPars}
\alias{setPars}
\title{Assign Values to Parameters}
\arguments{
\item{x}{A numeric vector or array, depending on the model's spatial
dimensions. Consult the help page of the sister method
\code{\link{setVars}} for details on the required input.}
}
\value{
\code{NULL} (invisible). The assigned numeric data are stored in the
  object and can be accessed by the \code{\link{getPars}} method.
}
\description{
Assign values to parameters of a \code{\link{rodeo}}-based model.
}
\note{
Look at the notes and examples for the \code{\link{setVars}} method.
}
\author{
\email{david.kneis@tu-dresden.de}
}
\seealso{
The corresponding 'get' method is \code{\link{getPars}}. Use
  \code{\link{setVars}} to assign values to variables rather than parameters.
  Consult the help page of the latter function for examples.
}

