% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\name{toymodel}
\alias{toymodel}
\title{Toy model}
\format{A single character string which can be used to save a toymodel
  specification to a file in JSON format (see example)}
\description{
A toy model in JSON notation
}
\examples{
\dontrun{
  data(toymodel)
  write(x=toymodel, file="toymodel.json")
  }
}

