% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\name{toymodel_pars}
\alias{toymodel_pars}
\title{Toy model parameters}
\format{A data frame with the following columns:
  \itemize{
    \item{name : }{Name of the parameter}
    \item{value : }{Value}
    \item{unit : }{Unit of the value}
    \item{description : }{Short description of the parameters's meaning}
  }}
\description{
Sample parameters for \code{toymodel}
}

